backend/requirements.txt
fastapi
uvicorn
pymysql
pydantic
python-dotenv
transformers
torch
streamlit
pandas
requests
plotly
streamlit_option_menu
